This is the first iteration of optimising the game for use with a screenreader.
When focussing an input element, it now announces the direction, the
clue and the number of characters if known.

The next iteration will also declare, which position in the respective
direction the element is in.

I've tested the changes using TalkBack on Android.

Example announcement:

"Direction across: Angry exclamation? You're looking at it! 9
characters. Direction down: Con men opt to reassemble part 9
characters."

The accessibility inspector is helpful. Try it on your own:
https://developer.mozilla.org/en-US/docs/Tools/Accessibility_inspector

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>